==> notebooks/modeling/06_model_table_tuning.ipynb <==
# Imports
import pandas as pd
import os
import numpy as np

# Read model_input
dtype_dict = {'HCP_ID': 'str',
    'PATIENT_AGE': 'Int64',
    'PATIENT_GENDER': 'category',
    'PATIENT_ID': 'str',
    'STATE': 'category',
    'HCP_SPECIALTY': 'category',
    'HCP_AGE': 'Int64',
    'HCP_GENDER': 'category',
    'STATE_NAME': 'category',
    'TXN_LOCATION_TYPE': 'category',
    'INSURANCE_TYPE': 'category',
    'TXN_TYPE': 'category',
    'TXN_DESC': 'category',
    'LOWCONT_COUNT': 'Int64',
    'MEDCONT_COUNT': 'Int64',
    'HIGHCONT_COUNT': 'Int64',
    'NUM_CONDITIONS': 'Int64',
    'NUM_CONTRAINDICATIONS': 'Int64',
    'NUM_SYMPTOMS': 'Int64',
    'DRUGA_COUNT': 'Int64',
    'TARGET': 'Int64'}

model_table = pd.read_csv('../../data/model_input/model_table.csv', dtype=dtype_dict)

# Parse dates to datetime
date_columns = ['TXN_DT', 'BIRTH_DT']
for col in date_columns:
    model_table[col] = pd.to_datetime(model_table[col], errors='coerce')


model_table.info()

model_table.drop(columns=['BIRTH_DT'], inplace=True)
model_table.drop(columns=['STATE_NAME'], inplace=True)
model_table.drop(columns=['PATIENT_AGE'], inplace=True)
model_table.drop(columns=['HCP_AGE'], inplace=True)
model_table.drop(columns=['DRUGA_COUNT'], inplace=True)

model_table.drop(columns=['TXN_TYPE'], inplace=True)
model_table.drop(columns=['TXN_DESC'], inplace=True)
model_table.drop(columns=['RANK'], inplace=True)


model_table

# Drop rows where 'HCP_ID' is NaN, non in target list
model_table = model_table.dropna(subset=['HCP_ID'])
model_table.info()

# Save Analytics Ready Dataset
model_table.to_csv('../../data/model_input/ads_model_table.csv', index=False)

model_table.info()


# Drop TARGET column
model_table = model_table.drop(columns=['TARGET'])

# Select 5 random rows
sampled_model_table = model_table.sample(n=5, random_state=1)

# Save to CSV
sampled_model_table.to_csv('../../data/model_input/file_to_predict.csv', index=False)
