==> notebooks/modeling/09_model_predict.ipynb <==
# Imports
import joblib
import pandas as pd

# Load the model
model = joblib.load('../../models/saved_models/random_forest_model.joblib')

# Define the correct datatypes for pandas
dtype_dict = {
    'HCP_ID': 'str',
    'PATIENT_AGE': 'Int64',
    'PATIENT_GENDER': 'category',
    'PATIENT_ID': 'str',
    'STATE': 'category',
    'HCP_SPECIALTY': 'category',
    'HCP_AGE': 'Int64',
    'HCP_GENDER': 'category',
    'STATE_NAME': 'category',
    'TXN_LOCATION_TYPE': 'category',
    'INSURANCE_TYPE': 'category',
    'TXN_TYPE': 'category',
    'TXN_DESC': 'category',
    'LOWCONT_COUNT': 'Int64',
    'MEDCONT_COUNT': 'Int64',
    'HIGHCONT_COUNT': 'Int64',
    'NUM_CONDITIONS': 'Int64',
    'NUM_CONTRAINDICATIONS': 'Int64',
    'NUM_SYMPTOMS': 'Int64',
    'DRUGA_COUNT': 'Int64',
    'TARGET': 'Int64'
}

input_data = pd.read_csv('../../data/model_input/file_to_predict.csv')


# Convert columns to the specified data types
for col, dtype in dtype_dict.items():
    if col in input_data.columns:
        if dtype == 'category':
            input_data[col] = input_data[col].astype('category')
        elif dtype == 'Int64':
            input_data[col] = pd.to_numeric(input_data[col], errors='coerce').astype('Int64')
        else:
            input_data[col] = input_data[col].astype(dtype)

# Parse dates to datetime
date_columns = ['TXN_DT']
for col in date_columns:
    if col in input_data.columns:
        input_data[col] = pd.to_datetime(input_data[col], errors='coerce')

# Display the DataFrame info
input_data.info()

input_data.head()

# Preprocess categorical features
categorical_columns = input_data.select_dtypes(include=['category']).columns
input_data = pd.get_dummies(input_data, columns=categorical_columns, drop_first=True)

# Drop unnecessary columns
input_data = input_data.drop(columns=['PATIENT_ID', 'TXN_DT', 'HCP_ID'])

input_data.info()

input_data.head()

# Ensure all expected columns are present
expected_columns = model.feature_names_in_
for col in expected_columns:
    if col not in input_data.columns:
        input_data[col] = False
        
#Reorder columns to match the training data
input_data = input_data[expected_columns]

input_data.info()

# Make prediction
prediction = model.predict(input_data)
#return {"prediction": prediction[0]}

prediction

df = pd.read_csv('../../data/model_input/file_to_predict.csv')
        
# Convert DataFrame to dictionary
data_dict = df.to_dict(orient='records')

import pandas as pd
from pydantic import BaseModel, create_model

# Generate the Pydantic model dynamically
fields = {col: (str, ...) if df[col].dtype == 'object' else (int, ...) for col in df.columns}
PatientData = create_model('PatientData', **fields)

# Print the dynamically created model schema
print(PatientData.schema_json(indent=2))

data_dict
